requirements.txt
google-generativeai
python-dotenv
Pillow
PyPDF2
flask
flask-cors
werkzeug
requests
